implement wallet scheme to wrap up the 2 known wallet models
this intends to simplify the implementations and expectations
for the different known wallet implemented by this library.

the idea is to keep an abstract model, generic enough yet simple to
represent what is the expected behaviour of a wallet in this library.
Then we implement the wallet for bip44 sequential address and the
wallet as implemented in daedalus.

Then, depending of the wallet, specific operations are added.