exe-common::sync: Hot-fix a u32 underflow panic
This was originally found on a first run of Hermes which would catch the
underflow on the genesis epoch. The finish_epoch() function would use an
imprecise logical expression to determine that the specified epoch has a
predecessor.

This commit rules fixes the underflow which would happen for
epoch_id equal to 0 a.k.a. genesis.

exe-common/src/sync.rs:
* Fix finish_epoch()'s assert condition