Add a function for verifying a block in isolation
I.e. this checks the internal consistency of the block (e.g. that
header proofs are consistent with the body, that the block signature
is correct, etc). It does not check whether the block is valid in the
context of the chain (e.g. that it doesn't spend already spent
outputs).