ChainState: Keep track of utxos
The new struct GenesisData (representing selected contents of the
genesis data file, e.g. mainnet-genesis.json) is used to initialize
the UTXO state from the AVVM distribution. Note that we don't parse
the genesis data file in this crate since we don't want a dependency
on a JSON parser.

ChainState::verify_block() updates the utxo state from every
transaction, removing spent txos and adding new txos. It also verifies
that the witnesses are authorized to spend the utxo (e.g. that the
utxo's address can be computed fromt the witness's public key).