Store chain state by (boundary) block hash rather than epoch ID
That is, chain state deltas are now stored in chainstate/<HASH> rather
than epoch/<N>/utxos, where <HASH> is the block hash of the boundary
block of the epoch. This ensures that we don't have to worry about
potential rollbacks that could invalidate a stored epoch.

While currently deltas are still between boundary blocks, the new
scheme makes it possible to store deltas between arbitrary blocks. For
example, we could ignore epochs entirely and store deltas for every
1000th block in the chain.

Also:

* Store more state on disk, in particular spent_txos and
  nr_transactions.

* spend_txos -> spent_txos