Add a function that returns the chain state at an arbitrary block
It seeks backwards through the chain until it reaches a block that has
chain state on disk. It then goes forward again to verify and apply
the missing blocks.

This function can be quite slow in debug mode because it has to
read/verify up to an epoch's worth of blocks, but in release mode it
only takes a few seconds. In the future we may want to write chain
state deltas more frequently than once per epoch though.