Crate network-core
Moved the abstract network interface definitions, incubated in
network-grpc, to a new dedicated crate network-core.

The interfaces rely on enum types generated from protobuf descriptions.
For this, the protobuf declarations of enums have been split out into
a dedicated package codes.proto which is used by the build script of
network-core to generate the types and make them available as
module codes.