Added new field `chain_state::ChainState::last_boundary_block_epoch` that stores optional epoch number for which this last boundary block were found, because after OBFT last EBB might be many epochs back.
In all places where `last_boundary_block` is re-assigned -  this new fields is assigned the epoch from the EBB "date". There's a piece of code that checks if new non-boundary block is from new epoch, which would indicate an OBFT epoch-transition, and in this case last EBB state potentially might be cleared, but for now - it's commented out.

In `verify_chain::ChainState::verify_block` updated the block where slot-leader is validated to first check if there's a known EBB for the current epoch, if so - slot-leader is validated as usual, but if not - OBFT validation is required, but for now yet unimplemented (TODO).