Fix LooseChainHeightEntry::header_hash()
HeaderHash and BlockHash are the same hash but just Blake2b256 vs raw byte slice of the same hash.
This caused header_hash() to re-hash the hash, giving incorrect hashes
and thus breaking the rollback code that it was written for.

This change was tested with the new rollback code and with a test
assert.